refactor(agent): extract tool execution into dedicated module
Move tool dispatch, parallelism, and delivery-default logic from
loop_.rs (5,973 → 5,595 lines) into agent/tool_execution.rs (469 lines).

Extracted:
- ToolExecutionOutcome, find_tool, execute_one_tool
- execute_tools_parallel, execute_tools_sequential
- should_execute_tools_in_parallel
- maybe_inject_channel_delivery_defaults
- Related unit tests (5 tests moved)

No public API changes. All existing tests pass.
